[SPARK-41246][CORE] Gate 64-bit RDD IDs behind spark.rdd.longIds.enabled; throw on Int overflow
In long-running Spark applications that create more than Int.MaxValue RDDs,
AtomicInteger.getAndIncrement() wraps to negative values.  The BlockId regex
"rdd_([0-9]+)_([0-9]+)" cannot parse negative IDs, producing
UnrecognizedBlockId exceptions and breaking the block cache.

Introduce a feature flag spark.rdd.longIds.enabled (default false).

When false (legacy mode):
- RDD IDs are still allocated from a 32-bit counter.
- If the counter overflows (reaches Int.MinValue), throw IllegalStateException
  with a message pointing to the flag, instead of silently producing negative IDs.

When true (long mode):
- RDD IDs are allocated from a 64-bit AtomicLong counter (no overflow risk).
- RDD.longId (Long) is the canonical identifier; RDD.id returns -1 as sentinel.
- persistentRdds, BlockInfoManager.rddToBlockIds, BlockManagerMessages.RemoveRdd,
  BlockManager/BlockManagerMaster.removeRdd, StorageUtils._rddBlocks/_rddStorageInfo,
  ContextCleaner.CleanRDD/CleanCheckpoint, SparkListenerUnpersistRDD.rddId, and
  JsonProtocol are all widened to Long.

Phase-1 compromises (fixed in subsequent PRs):
- RDDBlockId.rddId remains Int; the block cache is still limited to Int.MaxValue
  distinct RDD IDs even with the flag enabled.
- AppStatusListener.liveRDDs, ExecutorMonitor.cachedBlocks, and
  BasicEventFilterBuilder.liveRDDs are Int-keyed; operations for IDs > Int.MaxValue
  are silently skipped.
- The flag default will be flipped to true in PR 2 after RDDBlockId is widened.

- LongRddIdSuite (27 tests): systematic A/B/C/D scenario coverage for every
  touched component (SparkContext/RDD, persistentRdds, StorageUtils,
  SparkListenerUnpersistRDD, JsonProtocol, BasicEventFilter, ContextCleaner,
  BlockManagerMessages).
- SparkContextSuite (5 new tests): flag=true/false x in-range/overflow.
- BlockInfoManagerSuite (3 new tests): rddBlockIds Long key A/C/D.
- ExecutorMonitorSuite (3 new tests): onUnpersistRDD Long rddId A/C/D.
- ContextCleanerSuite: all 10 existing tests still pass.

+++ core/src/main/scala/org/apache/spark/scheduler/DAGScheduler.scala
@@ -249,7 +249,7 @@ private[spark] class DAGScheduler(
     if (injectShuffleFetchFailuresCleanerAttached) return
     sc.cleaner.foreach { cleaner =>
       cleaner.attachListener(new CleanerListener {
-        override def rddCleaned(rddId: Int): Unit = {}
+        override def rddCleaned(rddId: Long): Unit = {}
         override def shuffleCleaned(shuffleId: Int): Unit = {
           injectShuffleFetchFailuresCorruptedAttempt.remove(shuffleId)
           injectShuffleFetchFailuresPendingDelayedCorruption.remove(shuffleId)
